refactor : serve-static 모듈로 분리
application.js에 있는 정적파일 다루는 로직을 serve-static.js 파일로 분리
